requirements.txt
pandas
numpy
faker
scikit-learn
matplotlib
seaborn
nltk
textblob
xgboost
tensorflow
torch
transformers
faiss-cpu
sentence-transformers
streamlit
langchain
langchain_community
tf-keras
llama-cpp-python
flask
langchain_groq
datasets